Moved hierachyView_shouldNotChangeAfterScreenRotation to FormHierarchyTest
