+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/FormHierarchyTest.java
@@ -199,4 +199,19 @@ public class FormHierarchyTest {
                 .clickGoToArrow()
                 .checkIfElementInHierarchyMatchesToText("Group Name", 0);
     }
+
+    @Test
+    public void hierachyView_shouldNotChangeAfterScreenRotation() {
+        rule.startAtMainMenu()
+                .copyForm("repeat_group_form.xml")
+                .startBlankFormWithRepeatGroup("Repeat Group", "Grp1")
+                .clickOnDoNotAdd(new FormEntryPage("Repeat Group"))
+                .clickGoToArrow()
+                .clickGoUpIcon()
+                .checkIfElementInHierarchyMatchesToText("Group Name", 0)
+                .rotateToLandscape(new FormHierarchyPage("Repeat Group"))
+                .checkIfElementInHierarchyMatchesToText("Group Name", 0)
+                .rotateToPortrait(new FormHierarchyPage("Repeat Group"))
+                .checkIfElementInHierarchyMatchesToText("Group Name", 0);
+    }
 }
